another day at 42

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: hhakim <hhakim@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 19:29:53 by hhakim            #+#    #+#             */
-/*   Updated: 2022/11/02 16:44:32 by hhakim           ###   ########.fr       */
+/*   Updated: 2022/11/10 14:00:21 by hhakim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != 0)
